Update Page dto in orgs to accept Long values

+++ orgs/src/main/java/com/cryptopos/orgs/dto/Page.java
@@ -2,6 +2,6 @@ package com.cryptopos.orgs.dto;
 
 import java.util.Collection;
 
-public record Page<T>(int currentPage, int pageSize, int totalPages, Collection<T> data) {
+public record Page<T>(Long currentPage, Long pageSize, Long totalPages, Collection<T> data) {
 
 }
